+++ src/main/java/io/github/felixgilioli/felixutils/objects/ObjectUtils.java
@@ -1,0 +1,30 @@
+package io.github.felixgilioli.felixutils.objects;
+
+import java.util.Objects;
+import java.util.function.Predicate;
+
+/**
+ * Class with utility methods for working with {@link Object}.
+ */
+public class ObjectUtils {
+
+    private ObjectUtils() {}
+
+    /**
+     * Check that the specified object reference is not null and respect the condition
+     * @param obj the object reference to check for condition
+     * @return object if not null
+     * @throws IllegalArgumentException if object is does not respect the condition
+     * @throws NullPointerException if object is null
+     */
+    public static <T> T checkArgument(T obj, Predicate<T> condition) {
+        Objects.requireNonNull(obj);
+        Objects.requireNonNull(condition);
+
+        if (!condition.test(obj)) {
+            throw new IllegalArgumentException();
+        }
+
+        return obj;
+    }
+}