+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.1.1
+- [#2](https://github.com/felixgilioli/felix-utils/issues/2) - added ObjectUtils.checkArgument.
 - [#1](https://github.com/felixgilioli/felix-utils/issues/1) - Added method ListUtils.findFirst.
 
 ## 1.0.1